Fix the paths for admin-specific APIs and split UI client to user one and admin one
