+++ ui/src/MyPage.tsx
@@ -19,7 +19,8 @@ import {
 import { useHeader } from './Layout';
 import { useAuth } from './auth';
 import { Role } from './auth/role';
-import { Attendance, User, getUser, getUserAttendances, getUserAuth } from './client/http';
+import { Attendance, User } from './client/http/data';
+import { getUser, getUserAttendances, getUserAuth } from './client/http/default';
 import { toYYslashMMslashDDspaceHHcolonMM } from './common/date';
 import { useAdminMode } from './mode';
