+++ ui/src/SessionDetail/user/UserSessionDetail.tsx
@@ -2,7 +2,8 @@ import { useCallback, useEffect, useState } from 'react';
 import { useLocation, useNavigate, useParams } from 'react-router-dom';
 import { Container, Box, Button, CircularProgress } from '@mui/material';
 import { useHeader } from '../../Layout';
-import { Session, getSession } from '../../client/http';
+import { Session } from '../../client/http/data';
+import { getSession } from '../../client/http/default';
 import useHandleError from '../../common/error';
 import SessionInfo from '../common/SessionInfo';
 import SessionAttendanceTable from './AttendanceTable';
