+++ ui/src/client/http/default.ts
@@ -1,6 +1,7 @@
 import axios, { AxiosInstance } from 'axios';
 import { z } from 'zod';
-import { replaceCookieHeader, setAuthToken } from './auth';
+import { Attendance, AttendanceSchema, Session, SessionSchema, User, UserSchema } from './data';
+import setToken from './interceptor';
 
 const BASE_URL = import.meta.env.VITE_SERVER_ENDPOINT;
 
@@ -12,85 +13,7 @@ const client: AxiosInstance = axios.create({
   withCredentials: true,
 });
 
-client.interceptors.response.use((response) => {
-  const newToken = response.headers[replaceCookieHeader];
-  if (newToken) {
-    setAuthToken(newToken);
-  }
-  return response;
-});
-
-const UserSchema = z
-  .object({
-    id: z.string(),
-    name: z.string(),
-    generation: z.number(),
-    is_active: z.boolean(),
-    email: z.string(),
-    external_name: z.string(),
-  })
-  .transform((data) => ({
-    id: data.id,
-    name: data.name,
-    generation: data.generation,
-    isActive: data.is_active,
-    email: data.email,
-    externalName: data.external_name,
-  }));
-
-const AttendanceStatus = z.enum(['not_applied_yet', 'applied', 'ignored']);
-export const SessionAttendanceAppliedBy = z.enum(['unknown', 'unspecified', 'manual', 'form']);
-
-const SessionSchema = z
-  .object({
-    id: z.string(),
-    name: z.string(),
-    description: z.string(),
-    created_by: z.string(),
-    google_form_uri: z.string(),
-    google_form_id: z.string(),
-    created_at: z.string().transform((str) => new Date(str)),
-    starts_at: z.string().transform((str) => new Date(str)),
-    score: z.number(),
-    attendance_status: AttendanceStatus,
-    attendance_applied_by: SessionAttendanceAppliedBy,
-  })
-  .transform((data) => ({
-    ...data,
-    createdBy: data.created_by,
-    googleFormUri: data.google_form_uri,
-    googleFormId: data.google_form_id,
-    createdAt: data.created_at,
-    startsAt: data.starts_at,
-    attendanceStatus: data.attendance_status,
-    attendanceAppliedBy: data.attendance_applied_by,
-  }));
-
-const AttendanceSchema = z
-  .object({
-    id: z.string(),
-    session_id: z.string(),
-    session_name: z.string(),
-    session_score: z.number(),
-    session_started_at: z.string().transform((str) => new Date(str)),
-    user_id: z.string(),
-    user_external_name: z.string(),
-    user_generation: z.number(),
-    user_joined_at: z.string().transform((str) => new Date(str)),
-    created_at: z.string().transform((str) => new Date(str)),
-  })
-  .transform((data) => ({
-    id: data.id,
-    sessionId: data.session_id,
-    sessionName: data.session_name,
-    sessionScore: data.session_score,
-    sessionStartedAt: data.session_started_at,
-    userId: data.user_id,
-    userExternalName: data.user_external_name,
-    userGeneration: data.user_generation,
-    userJoinedAt: data.user_joined_at,
-    createdAt: data.created_at,
-  }));
+client.interceptors.response.use(setToken);
 
 const GetUserAttendancesResponseSchema = z.object({
   attendances: z.array(AttendanceSchema),
@@ -99,10 +22,6 @@ const GetUserAttendancesResponseSchema = z.object({
 const GetSessionAttendancesResponseSchema = z.object({
   attendances: z.array(AttendanceSchema),
 });
-
-export type User = z.infer<typeof UserSchema>;
-export type Session = z.infer<typeof SessionSchema>;
-export type Attendance = z.infer<typeof AttendanceSchema>;
 
 const ListUsersResponseSchema = z
   .object({
@@ -117,7 +36,6 @@ const ListUsersResponseSchema = z
   }));
 const UserResponseSchema = UserSchema;
 
-const SessionsResponseSchema = z.array(SessionSchema);
 const SessionResponseSchema = SessionSchema;
 
 const ListSessionsResponseSchema = z
@@ -144,24 +62,9 @@ export const getUser = async (id: string): Promise<User> => {
   return UserResponseSchema.parse(response.data);
 };
 
-export const createUser = async (name: string, generation: string, isActive: boolean, email: string): Promise<User> => {
-  const response = await client.post('/users', {
-    name,
-    generation,
-    is_active: isActive,
-    email,
-  });
-  return UserResponseSchema.parse(response.data);
-};
-
 export const getSession = async (id: string): Promise<Session> => {
   const response = await client.get(`/sessions/${id}`);
   return SessionResponseSchema.parse(response.data);
-};
-
-export const getSessions = async (): Promise<Session[]> => {
-  const response = await client.get('/sessions');
-  return SessionsResponseSchema.parse(response.data);
 };
 
 export type ListSessionsResponse = z.infer<typeof ListSessionsResponseSchema>;
@@ -169,30 +72,6 @@ export type ListSessionsResponse = z.infer<typeof ListSessionsResponseSchema>;
 export const listSessions = async (offset: number, pageSize: number): Promise<ListSessionsResponse> => {
   const response = await client.get('/sessions', { params: { offset, pageSize } });
   return ListSessionsResponseSchema.parse(response.data);
-};
-
-export const createSession = async (
-  name: string,
-  description: string,
-  startsAt: Date,
-  score: number,
-): Promise<Session> => {
-  const response = await client.post('/sessions', {
-    name,
-    description,
-    starts_at: startsAt.toISOString(),
-    score,
-  });
-  return response.data.id;
-};
-
-export const deleteSession = async (id: string): Promise<void> => {
-  await client.delete(`/sessions/${id}`);
-};
-
-export const createSessionForm = async (sessionId: string): Promise<string> => {
-  const response = await client.post(`/sessions/${sessionId}/attendance-form`);
-  return response.data.form_url;
 };
 
 export const getUserAttendances = async (userId: string): Promise<Attendance[]> => {
@@ -203,10 +82,6 @@ export const getUserAttendances = async (userId: string): Promise<Attendance[]> 
 export const getSessionAttendances = async (sessionId: string): Promise<Attendance[]> => {
   const response = await client.get(`/sessions/${sessionId}/attendances`);
   return GetSessionAttendancesResponseSchema.parse(response.data).attendances;
-};
-
-export const markUsersAsPresent = async (sessionId: string, userIds: string[]): Promise<void> => {
-  await client.post(`/sessions/${sessionId}/attendance/manual`, { user_ids: userIds });
 };
 
 const GetHalfYearAttendancesResponseSchema = z
